Fix for #1801. (#1803)
Three test cases had compilation errors as the Ids of elements had changed on the activity. This change was not incorporated in the test cases.

Updated the files to have correct UI element Ids

p.s There are warnings dur to deprecated FLAGS eg. FLAG_TURN_SCREEN_ON. Did not change it to not break for anyone using API level before 27